Bomb Squad
Projct was developed for the purpose of GUSAC Event. This is a quiz game
where the question has a 90 sec time limit and the user has to answer
the question to diffuse the bomb. The user will have 5 lives. The bomb
will be exploded if the answer is wrong or if he runs out of time.